Add --snapcraft-deb and --snapcraft-snap options
Teach vm_setup to install either the snapcraft deb or the snapcraft snap
depending on the option given.

They differ on behaviour and release management. We're having trouble
aligning our CI with "official" builds on Launchpad. We've been advised
to try the edge snap which has some non-determinism removed. And
we're told that Launchpad is expected to switch to the snap eventually.
So permit both so that we can tweak our CI accordingly.