bootspeed: retry to `snap refresh` if the change status is 'Error'
This is an attempt to make the script more robust when `snap refresh`
fails. In this case the change status listed in `snap changes` is
'Error', and when this happens the script will now run another `snap
refresh` in the provisioned system.

The error condition is not easy to reproduce, so this workaround has not
been tested in an actual error case. Time will tell if it's effective.