measure-device.sh: bump the testflinger timeout to 16h
Queues for devices are sometimes very long, let's bump the timeout.
We currently run 1job/day/device, so waiting for 16h is both generous
and safe from overlaps.

This could be handeld as a Jenkins timeout (and we do have one), but
handling the timeout in the script is better as the job is removed from
the queue before exiting, something Jenkins can't easily do.