measure-device: use BUILD_TIMEOUT if Jenkins exported it
If the Jenkins job has a timeout and exported a BUILD_TIMEOUT
environment variable, allocate most of the time the job is allowed to
run for the testflinger run. Testflinger is called using timeout(1),
which exits with RC=124 if the timeout is reached. The RC can be used
by Jenkins to set the build status to "unstable" in case of a
testflinger timeout.