copr_build: allow to specify the copr project name via --project
Up to this commit copr_build supports two copr projects:

 - cloud-init (default)
 - cloud-init-dev (if --dev is specified).

The project_id of the two projects is hardcoded.

This change supersedes this adding a --project flag that allows to
specify a project name on the command line. It is assumed that the
projects reside under the '@cloud-init' group. The default project
remains 'cloud-init'.

The --dev flag has been removed, requiring scripts that build against
cloud-init-dev to be updated:

  ./copr_build --dev

becomes:

  ./copr_build --project cloud-init-dev

copr_build verifies that the build succeeded against a list of chroots.
This used to be hardcoded as ['epel-8-x86_64']. The default list is now
empty; the list chroots to verify can be set via --test-chroot CHROOT,
which can be specified multiple times.

These changes are needed to build the stable-19.4 branch on centos/7.
A separate project is needed as the build chroots are specified per
project.