copr_build: drop --srpm; make srpm a positional argument
Drop the --srpm option and make the srpm file a positional argument.

Required options are discouraged by the argparse documentation:

    Required options are generally considered bad form because
    users expect options to be optional, and thus they should
    be avoided when possible.

Moreover required options are still listed under 'optional arguments'
in the --help printout, which is confusing.

The introduction of --project already changes the copr_build syntax,
so this is a good moment to also introduce this other breaking change.