LMA: Implement Dockerfile and reorganize test files
This commit implements a simple Dockerfile which creates an image that
contains everything that is needed to run the LMA integration tests.

It also creates an entrypoint script that will be invoked when the
container runs, and moves the global variable definitions from the
helper.sh script to a globals.sh one.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>