LMA: Update docker-compose.yml and the run-tests script
Now that the tests are running inside a container, we have to update
docker-compose.yml and make it actually fire up the "tests" container
using our image.

This commit also updates the "run-tests" script in order to make it
properly invoke docker-compose, display the results and exit with the
correct exit code.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>