OCI unit tests: Use psql from container
This commit is based on Andrea's suggestion from PR #60.  Instead of
relying on the host's psql binary, we can (and indeed should!) use the
container's psql.  This has a bunch of positive points:

- It makes it unnecessary to install test dependencies on the host.

- It exercises the psql from inside the container, which is software
  that we will ship to the users anyway.

- It uses non-localhost hostnames for the tests (albeit we'd still
  be using a local docker network bridge).

I've made sure the tests are still passing, and I don't have psql
installed in the host machine.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>