OCI unit test: mysql
This commit implements the unit test for mysql.  It is loosely based
on the postresql unit test, but has a few important modifications:

- It was necessary to extend the 'wait_container_ready' helper
  function and make it accept an optional argument specifying the
  timeout.  mysqld takes a long time to start.

- I took the opportunity and also extended 'stop_container_sync' to
  take an optional timeout as its second argument, although it's not
  being used so far.

- The unit test was written using the same idiom I used for the redis
  unit test (with functions that encapsulate the execution of both the
  server and CLI programs).

Because the mysql server takes a long time to start, running all of
the tests will also be time-consuming.  I'm also not entirely
confident that a 60-second timeout to start the server will be enough
in all scenarios, but we can always increase it if needed.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>