Add try except block when deleting ec2 resources
Currently, we do not check for failures when deleting
ec2 resources. This means that we will abort the cleanup
script if we fail to delete any resource. This is not ideal,
since we should be trying to cleanup as much as we can on each
run. We are updating the script to do that