OCI: Fix apache2 configure loading
The apache2 unit test is using $PWD when composing the path for the
configure files to be mounted inside the container.  This doesn't work
when we're running these tests from an external directory.  We should
use the same pattern we're using to source the helper scripts here,
with the addition of calling "realpath", which is necessary because
docker requires that mount paths be fullpaths.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>