OCI: Implement test to check if images have manifest files
Given the requirement that the images must contain manifest
files (under /usr/share/rocks) for the Security team, it's a good idea
to extend our unit tests and make sure we confirm the presence of said
files.

I'd have preferred something a little less repetitive than this
change, but unfortunately the way we start each service is different,
so we can't generalize it.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>