OCI: Remove previous image before starting test
When running a test whose image cannot be downloaded from the
registry, docker will use the cached image (from a previous run).  In
order to prevent this misleading scenario from happening, this commit
forces the removal of an existing image before we start testing.

Note that we just remove the image if "${DOCKER_IMAGE}" (which
contains the full image spec) starts with "${DOCKER_REGISTRY}".  This
is necessary in order to continue supporting testing locally built
images.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>