OCI: Use "classic" interface when querying prometheus' endpoints
The version of prometheus that's being delivered in our OCI image has
updated their web interface and are now making extensive use of
JavaScript; it actually *requires* JavaScript now, so a simple "curl"
won't be able to obtain useful information from the pages.

Our test was written for a previous version of prometheus, which
worked fine with "curl".  Fortunately, upstream has kept the "classic"
interface around, so it's still possible to query the pages we need
using it.

Of course, there's the question of how long upstream will be willing
to maintain the classic interface.  We might have to revisit this
later, but for now it works.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>