OCI: Skip testing with prom/pushgateway on non-amd64 arches
The prom/pushgateway image is not published for all architectures that
are supported by us.  For this reason, when we run the test on those
architectures (for example, s390x), it fails.

In order to avoid such failures, this commit implements skipping of
tests that use the prom/pushgateway image on non-amd64 architectures.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>