OCI: Adjust message when number of images differ between registries
Currently we print a message tagged as "error" when there is a
mismatch between the number of images available in one registry and
another.  We also print a misguided message saying that the test will
be aborted, when in fact it is not.

This message can be just a warning, since it doesn't actually impact
the actual test.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>