OCI: Check whether the manifest file(s) is(are) empty
Athos has recently worked on (and fixed) a bug on the golang manifest
generator script which rendered empty manifest files for some images.
Apparently this bug has silently crept in some time ago, and we didn't
notice.  I thought that it'd be useful to expand our current
"is-the-manifest-file-present" unit test and make it also error if the
manifest file is empty (which should never happen).

I wasn't entirely sure whether this should be a new test or not, but I
think it makes sense (and is cleaner) to do it in the existing test.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>