OCI: Update docker-compose.yml and always build the test image
This commit serves two purposes.

First, it updates the docker-compose.yml file and makes it pull the
OCI images from the "ubuntu" namespace, and not the (retired)
squeakywheel one.  Note that the cortex image doesn't need to be
provided with a configuration file anymore, so this commit also gets
rid of this file.

Second, it changes how we generate/maintain the "lma-integration-test"
image.  We used to build it locally and upload it to the squeakywheel
namespace, but that's not really necessary.  In order to prevent
namespace pollution, we now always build the test image locally and
use it.  Although this is a bit more resource-intensive than the
previous approach, it shouldn't really impact the jenkins node because
the image is really small and builds very quick.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>