Use local cqlsh image as a fallback only
Use the official Cassandra image as a source for the cqlsh CLI
application. In case the image is not available for a given platform, we
fallback to building a cqlsh image locally. This is a compromise between
saving CI resources when possible and still ensuring we will have an
image with cqlsh to test Cassandra.

Signed-off-by: Athos Ribeiro <athos.ribeiro@canonical.com>