OCI: Consolidate $ROOTDIR usage (and fix a few pending issues)
I was testing the new telegraf OCI image for 21.10 from inside the
image's repository (i.e., I was invoking the unittest script while
outside its working directory), and noticed that the test was failing
because it was using $PWD indiscriminately.

I took the opportunity to improve the handling of these scenarios.  We
already had a $ROOTDIR variable being declared in the test_helper.sh
script, but it contained the wrong value and wasn't being used by
anyone.  I fixed it and converted the existing calls to $(dirname
"$0") and made them use $ROOTDIR instead.  While at it, I have also
removed the needless $PATH manipulation that was being done in the
same script.

Signed-off-by: Sergio Durigan Junior <sergio.durigan@canonical.com>