metric: rename observe-only -> server-simple
Now that we install workloads it is no more "observer-only".
The attribute the current set shares is to be simple to gather and
understand (as selected as subset out of US013).

Furthermore this was the only job that had no team designation in the
name, let us add that as well.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>