metric-server-simple: gather service status
Sometimes a rare event might happen that is hard to recerate and debug
later. So far we have seen two times that service startup was bad.

Therefore it would be very helpful to have at least the general final
status of all services in the log as it might indicate what was broken.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>