metric-server-simple: do not trace flaky cache values
/var/cache/apt can fluctuate a lot depending on how many packages
might need to be updated as (relative to the daily image) to install
packages of the workload. We are not interested in those fluctuations
as they would auto-clean over time anyway.

Therefore run `apt clean` after installing packages for the loaded
stage, but we want to know the initial image size if e.g. a clean was
forgotten on build, so do not clean it there.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>