metric-server-simple: increase VM size
Since the changes to the mantic minimal image this OOM crashes.
On one hand that increase was worth to know about, but on the other
hand all the data we get since then is flaky at beast.

While this change itself might shift values a bit, they will at least
be more reliable again. The intention is to consider reverting  this
change once [1][2] are resolved.

[1]: https://warthogs.atlassian.net/browse/CPC-2985
[2]: https://bugs.launchpad.net/cloud-images/+bug/2032933

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>