scripts: add cloud-init QEMU bootspeed script
In QEMU, launch daily cloudimages and cloudimages with an upgraded
cloud-init package to compare boot speed metrics.

Launch --number-of-samples in QEMU of both the control(clean daily)
and the upgraded image. Harvest any bootspeed artifacts from
systemd-analyze, cloud-init analyze, journalctl and logs for each
instance launched.

Average the boottimes per relevant systemd service and compare
control image to upgraded image to generate a report of
significant time deltas in systemd services between the two images.

This script is intended to be run on a system with sudo access
as it uses use mount-image-callback to create mounts of the base ISO
to installe the cloud-init package from a PPA or a local deb file
on top of the control ISO.