ha: add initial pgsql resource agent test
This tests a simple master/slave replication setup as described in the
upstream documentation in [1].

First, it sets up an PostgreSQL HA cluster with 2 nodes in a
master/slave configuration. Then, it simulates a failure in the master
node to verify if the slave is successfully promoted.

This does not use a floating IP address for service access since it
would depend on another resource agent.

While shunit2 is a xUnit based framework, it does execute the tests in
the order they are defined by default. This test script does rely on the
order each test is called (i.e., the tests are not independent). This
was done this way to avoid resetting the cluster state for each test we
want to perform since this would require more resources for our runs.

[1] https://wiki.clusterlabs.org/wiki/PgSQL_Replicated_Cluster